Add trail, trailLeft and trailRight to Grid to capture player trail and resulting polygons.
The player will create 2 polygons with a successful trail. The
trailLeft and trailRight members will capture the polygon points.

The trail points are kept separate for cases neither of the two
polygons can be captured/filled - balls inside the polygons - and only
the trail can be filled.

Game Rule: the player is not allowed to back on its trail - loss of
life.